Move go to flowr button to the right
