+++ src/wexond/renderer/app/constants/icons.ts
@@ -14,6 +14,7 @@ import bookmarks from '~/shared/resources/icons/bookmarks.svg';
 import settings from '~/shared/resources/icons/settings.svg';
 import extensions from '~/shared/resources/icons/extensions.svg';
 import window from '~/shared/resources/icons/window.svg';
+import home from '~/shared/resources/icons/home.svg';
 import more from '~/shared/resources/icons/more.svg';
 import find from '~/shared/resources/icons/find.svg';
 import edit from '~/shared/resources/icons/edit.svg';
@@ -47,6 +48,7 @@ export const icons = {
   settings,
   extensions,
   window,
+  home,
   more,
   find,
   edit,
